update Descartes List: add confirmation before delete, refine delete and update logic, link titles to details view, and enhance table styles and routing

+++ src/app/descartes-square/descartes-list/descartes-list.ts
@@ -5,13 +5,15 @@ import {MoreOptions} from '@core/components/more-options/more-options';
 import {IMoreOptions} from '@core/interfaces/more-options.interface';
 import {MatIcon} from '@core/enums/mat-icon.enum';
 import {MoreOptionAction} from '@core/enums/more-options-action.enum';
-import {Router} from '@angular/router';
+import {Router, RouterLink} from '@angular/router';
 import {IDescartesSolution} from '../definitions/interfaces/descartes-solution.interface';
 import {MatButton} from '@angular/material/button';
+import {filter, first, tap} from 'rxjs';
+import {ConfirmService} from '@core/services/confirm.service';
 
 @Component({
   selector: 'app-descartes-list',
-  imports: [MatTableModule, MoreOptions, MatButton],
+  imports: [MatTableModule, MoreOptions, MatButton, RouterLink],
   templateUrl: './descartes-list.html',
   styleUrl: './descartes-list.scss'
 })
@@ -27,7 +29,11 @@ export class DescartesList implements OnInit {
       text: 'Delete',
       action: MoreOptionAction.Delete
     }]
+
+  readonly #confirmService = inject(ConfirmService);
+
   dataSource = signal<IDescartesSolution[]>([]);
+
   displayedColumns = ['id', 'title', 'conclusion', 'options'];
 
   #router = inject(Router);
@@ -39,18 +45,24 @@ export class DescartesList implements OnInit {
 
 
   update(id: string): void {
-    this.#router.navigate(['descartes-square/edit', id]);
+    this.#router.navigate([`descartes-square/list/${id}/edit`]);
   }
 
   delete(id: string): void {
-    const currList: IDescartesSolution[] = JSON.parse(localStorage.getItem(LocalStorageKeys.LIST) || '[]')
-      .filter((item: IDescartesSolution) => item.id !== id);
-    localStorage.setItem(
-      LocalStorageKeys.LIST,
-      JSON.stringify(currList)
-    )
+    this.#confirmService.confirm('Are you sure you want to delete this record?')
+      .pipe(
+        first(),
+        filter(Boolean),
+        tap(() => {
+          const currList: IDescartesSolution[] = JSON.parse(localStorage.getItem(LocalStorageKeys.LIST) || '[]')
+            .filter((item: IDescartesSolution) => item.id !== id);
+          localStorage.setItem(
+            LocalStorageKeys.LIST,
+            JSON.stringify(currList)
+          )
 
-    this.dataSource.set(currList);
+          this.dataSource.set(currList);
+        })).subscribe()
   }
 
   addData(): void {
